+++ client/src/game/shapes/templates.ts
@@ -1,4 +1,5 @@
 import type { ApiRectShape, ApiShape } from "../../apiTypes";
+import type { ShapeOptions } from "../models/shapes";
 import { BaseAuraStrings, BaseTemplateStrings, BaseTrackerStrings, getTemplateKeys } from "../models/templates";
 import type { BaseAuraTemplate, BaseTemplate, BaseTrackerTemplate } from "../models/templates";
 import { aurasToServer } from "../systems/auras/conversion";
@@ -66,6 +67,19 @@ export function toTemplate(shape: ApiShape): BaseTemplate {
     // should be template[key], but this is something that TS cannot correctly infer (issue #31445)
     // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment
     for (const key of BaseTemplateStrings) (template as any)[key] = shape[key];
+
+    // We usually don't want to save the options that are currently stored on the shape
+    // there are some options however that SHOULD transfer. (currently only notes)
+    if (shape.options) {
+        const options: Partial<ShapeOptions> = {};
+        const { templateNoteIds } = Object.fromEntries(JSON.parse(shape.options) as any[]) as Partial<ShapeOptions>;
+        if (templateNoteIds) {
+            options.templateNoteIds = templateNoteIds;
+        }
+        if (Object.keys(options).length > 0) {
+            template.options = JSON.stringify(Object.entries(options));
+        }
+    }
 
     template.auras = [];
     template.trackers = [];