+++ CHANGELOG.md
@@ -68,6 +68,7 @@ tech changes will usually be stripped from release notes for the public
     -   Fix 'add shape' and 'remove shape' events not being synced immediately if you only have view access
     -   Note icon on shape was drawn in strange locations for shapes larger than 1x1
     -   Fix shape removal not properly removing the shape from related notes client-side
+    -   Fix notes not being stored in shape templates
 -   Groups:
     -   The 'edit shape' groups tab was completely broken, this has been resolved
     -   Multiple things in the groups tab have become more responsive to changes